Created Blob storage for storing and uploading pictures.
For this commit, I created an Azure Blob Storage and connected it to the Database, so now a user can easily place his/her new profile picture and can be seen by everyone , so the company can see him. I added more fields to the Company model so it can attract more users to candidate for their job listings.